# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Orange Money",
        "description": "API complète pour le système de paiement Orange Money avec authentification OTP et mot de passe",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/comptes": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte Orange Money",
                "operationId": "d46221922144c672a4a500c7f3755629",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_client",
                                    "numero_compte",
                                    "nom",
                                    "email",
                                    "telephone",
                                    "password"
                                ],
                                "properties": {
                                    "id_client": {
                                        "type": "string",
                                        "example": "CLI001"
                                    },
                                    "numero_compte": {
                                        "type": "string",
                                        "example": "OM001234567890"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "77 123 45 67"
                                    },
                                    "type_compte": {
                                        "type": "string",
                                        "enum": [
                                            "courant",
                                            "epargne",
                                            "entreprise"
                                        ],
                                        "example": "courant"
                                    },
                                    "statut_compte": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "inactif",
                                            "bloque",
                                            "suspendu"
                                        ],
                                        "example": "actif"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompteResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/compte/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir le solde du compte connecté",
                "operationId": "2b6a0439d95322a21d781e536f5f8e04",
                "responses": {
                    "200": {
                        "description": "Solde du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "solde": {
                                            "description": "Solde calculé du compte",
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompteResource": {
                "title": "Compte Resource",
                "description": "Ressource représentant un compte Orange Money formaté",
                "properties": {
                    "id": {
                        "description": "ID du compte",
                        "type": "integer"
                    },
                    "id_client": {
                        "description": "ID unique du client",
                        "type": "string"
                    },
                    "numero_compte": {
                        "description": "Numéro unique du compte",
                        "type": "string"
                    },
                    "nom": {
                        "description": "Nom du compte",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email du compte",
                        "type": "string",
                        "format": "email"
                    },
                    "telephone": {
                        "description": "Numéro de téléphone",
                        "type": "string"
                    },
                    "type_compte": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "courant",
                            "epargne",
                            "entreprise"
                        ]
                    },
                    "statut_compte": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "inactif",
                            "bloque",
                            "suspendu"
                        ]
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token d'accès Bearer obtenu via OTP ou authentification par mot de passe",
                "scheme": "bearer"
            }
        }
    },
...
}